Always update last_fixed[cpu]
We don't care if last_fixed[cpu] was actually fixed, we just need to know that
last_fixed[cpu] was the last page fault on "cpu", so we update it early to
minimize the number of not cleared pages.

Also small updates on moca script help function.

Perf test => do not compress traces