Re enable false pf and fix start bug
False pagefault were disabled by a bad test !pte_presen(*pte) instead of
pte_present(*pte) when clearing flags ...

By re enabling false page faults, a bug occurs at the beginning of the
execution, it seems to happens if a page is marked not present during the
initialisation of a task. To fix this, the first time that the monitor thread
see a task it touch it and do nothing. False page fault will only be added the
second time it saw the task.